move query to query folder
